tests/wpt/tests/svg/painting/parsing/fill-opacity-invalid.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="800px">
  <title>SVG Painting: parsing fill-opacity with invalid values</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/painting.html#FillOpacityProperty"/>
    <h:meta name="assert" content="fill-opacity supports only the grammar 'alpha-value'."/>
  </metadata>
  <g id="target"></g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/parsing-testcommon.js"/>
  <script><![CDATA[

test_invalid_value("fill-opacity", "1.");
test_invalid_value("fill-opacity", "2 3");

  ]]></script>
</svg>
